+++ src/basic/__tests__/basic.test.tsx
@@ -8,13 +8,12 @@ import {
   screen,
   within,
 } from "@testing-library/react";
-import CartPage from "@/refactoring/pages/cart";
-import AdminPage from "@/refactoring/pages/admin";
-import { CartItem } from "@/refactoring/features/cart/model";
-import { Coupon } from "@/refactoring/entities/coupon/model";
-import { Product } from "../../refactoring/entities/product/model/product";
-import { useCart, useCoupons, useProducts } from "../../refactoring/hooks";
-import * as cartUtils from "../../refactoring/models/cart";
+import { CartPage } from "@/refactoring/pages/cart";
+import { AdminPage } from "@/refactoring/pages/admin";
+import { CartItem } from "@/refactoring/features/cart";
+import { Coupon } from "@/refactoring/entities/coupon";
+import { Product } from "@/refactoring/entities/product";
+import { useCart, useCoupons, useProducts } from "@/refactoring/hooks";
 import * as model from "@/refactoring/features/cart/model";
 
 const mockProducts: Product[] = [
