+++ src/origin/components/CartPage.tsx
@@ -3,6 +3,7 @@ import { CartItem } from "@/refactoring/features/cart";
 import { Coupon } from "@/refactoring/entities/coupon";
 import { Product, ProductInfoCard } from "@/refactoring/entities/product";
 import { AddToCartButton } from "@/refactoring/features/cart";
+import { ManageCartItemCard } from "@/refactoring/features/cart/ui";
 
 interface Props {
   products: Product[];
@@ -32,13 +33,13 @@ export const CartPage = ({ products, coupons }: Props) => {
     });
   };
 
-  const removeFromCart = (productId: string) => {
+  const handleRemoveFromCart = (productId: string) => {
     setCart((prevCart) =>
       prevCart.filter((item) => item.product.id !== productId)
     );
   };
 
-  const updateQuantity = (productId: string, newQuantity: number) => {
+  const handleUpdateQuantity = (productId: string, newQuantity: number) => {
     setCart((prevCart) =>
       prevCart
         .map((item) => {
@@ -106,18 +107,6 @@ export const CartPage = ({ products, coupons }: Props) => {
   const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } =
     calculateTotal();
 
-  const getAppliedDiscount = (item: CartItem) => {
-    const { discounts } = item.product;
-    const { quantity } = item;
-    let appliedDiscount = 0;
-    for (const discount of discounts) {
-      if (quantity >= discount.quantity) {
-        appliedDiscount = Math.max(appliedDiscount, discount.rate);
-      }
-    }
-    return appliedDiscount;
-  };
-
   const applyCoupon = (coupon: Coupon) => {
     setSelectedCoupon(coupon);
   };
@@ -133,6 +122,7 @@ export const CartPage = ({ products, coupons }: Props) => {
               const remainingStock = getRemainingStock(product);
               return (
                 <ProductInfoCard
+                  key={product.id}
                   product={product}
                   remainingStock={remainingStock}
                   footer={
@@ -151,49 +141,13 @@ export const CartPage = ({ products, coupons }: Props) => {
 
           <div className="space-y-2">
             {cart.map((item) => {
-              const appliedDiscount = getAppliedDiscount(item);
               return (
-                <div
+                <ManageCartItemCard
                   key={item.product.id}
-                  className="flex justify-between items-center bg-white p-3 rounded shadow"
-                >
-                  <div>
-                    <span className="font-semibold">{item.product.name}</span>
-                    <br />
-                    <span className="text-sm text-gray-600">
-                      {item.product.price}원 x {item.quantity}
-                      {appliedDiscount > 0 && (
-                        <span className="text-green-600 ml-1">
-                          ({(appliedDiscount * 100).toFixed(0)}% 할인 적용)
-                        </span>
-                      )}
-                    </span>
-                  </div>
-                  <div>
-                    <button
-                      onClick={() =>
-                        updateQuantity(item.product.id, item.quantity - 1)
-                      }
-                      className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                    >
-                      -
-                    </button>
-                    <button
-                      onClick={() =>
-                        updateQuantity(item.product.id, item.quantity + 1)
-                      }
-                      className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                    >
-                      +
-                    </button>
-                    <button
-                      onClick={() => removeFromCart(item.product.id)}
-                      className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
-                    >
-                      삭제
-                    </button>
-                  </div>
-                </div>
+                  item={item}
+                  onChangeQuantity={handleUpdateQuantity}
+                  onRemove={handleRemoveFromCart}
+                />
               );
             })}
           </div>
