+++ src/refactoring/features/cart/model.ts
@@ -151,3 +151,15 @@ export const findCartItemByProductId = (
 ) => {
   return cart.find((item) => item.product.id === productId);
 };
+
+export const getAppliedDiscount = (item: CartItem) => {
+  const { discounts } = item.product;
+  const { quantity } = item;
+  let appliedDiscount = 0;
+  for (const discount of discounts) {
+    if (quantity >= discount.quantity) {
+      appliedDiscount = Math.max(appliedDiscount, discount.rate);
+    }
+  }
+  return appliedDiscount;
+};
