+++ src/refactoring/features/cart/ui/index.ts
@@ -1,1 +1,2 @@
 export { default as AddToCartButton } from "./AddToCartButton";
+export { default as ManageCartItemCard } from "./ManageCartItemCard";
